+++ src/typeorm-bypasses/polaris-entity-manager.ts
@@ -4,17 +4,25 @@ import {
     DeepPartial,
     DeleteResult,
     EntityManager,
+    EntitySchema,
     FindOneOptions,
+    In,
     UpdateResult,
 } from 'typeorm';
-import { PolarisCriteria } from './contextable-options/polaris-criteria';
-import { PolarisFindManyOptions } from './contextable-options/polaris-find-many-options';
-import { PolarisFindOneOptions } from './contextable-options/polaris-find-one-options';
-import { PolarisSaveOptions } from './contextable-options/polaris-save-options';
-import { DataVersionHandler } from './handlers/data-version-handler';
-import { FindHandler } from './handlers/find-handler';
-import { SoftDeleteHandler } from './handlers/soft-delete-handler';
-import { CommonModel } from './models/common-model';
+import { RepositoryNotFoundError } from 'typeorm/error/RepositoryNotFoundError';
+import {
+    CommonModel,
+    PolarisCriteria,
+    PolarisFindManyOptions,
+    PolarisFindOneOptions,
+    PolarisSaveOptions,
+} from '..';
+import { DataVersionHandler } from '../handlers/data-version-handler';
+import { FindHandler } from '../handlers/find-handler';
+import { SoftDeleteHandler } from '../handlers/soft-delete-handler';
+import { PolarisConnection } from './polaris-connection';
+import { PolarisRepository } from './polaris-repository';
+import { PolarisRepositoryFactory } from './polaris-repository-factory';
 
 export class PolarisEntityManager extends EntityManager {
     private static async setInfoOfCommonModel(
@@ -23,36 +31,75 @@ export class PolarisEntityManager extends EntityManager {
     ) {
         if (maybeEntityOrOptions instanceof Array) {
             for (const t of maybeEntityOrOptions) {
-                t.dataVersion = context.returnedExtensions.globalDataVersion;
-                t.realityId = (context.requestHeaders && context.requestHeaders.realityId) || 0;
+                t.dataVersion = context?.returnedExtensions?.globalDataVersion;
+                t.realityId = context?.requestHeaders?.realityId ?? 0;
                 PolarisEntityManager.setUpnOfEntity(t, context);
             }
-        } else {
-            maybeEntityOrOptions.dataVersion = context.returnedExtensions.globalDataVersion;
-            maybeEntityOrOptions.realityId =
-                (context.requestHeaders && context.requestHeaders.realityId) || 0;
+        } else if (maybeEntityOrOptions instanceof Object) {
+            maybeEntityOrOptions.dataVersion = context?.returnedExtensions?.globalDataVersion;
+            maybeEntityOrOptions.realityId = context?.requestHeaders?.realityId ?? 0;
             PolarisEntityManager.setUpnOfEntity(maybeEntityOrOptions, context);
         }
     }
 
-    private static setUpnOfEntity(entity: any, context: any) {
-        if (context.requestHeaders) {
-            const id = context.requestHeaders.upn || context.requestHeaders.requestingSystemId;
-            if (entity.lastUpdateTime == null) {
-                entity.createdBy = id;
+    private static setUpnOfEntity(entity: any, context: PolarisGraphQLContext) {
+        if (context?.requestHeaders) {
+            if (entity.creationTime === undefined) {
+                entity.createdBy =
+                    context?.requestHeaders?.upn || context?.requestHeaders?.requestingSystemId;
+                entity.lastUpdatedBy = entity.createdBy;
+            } else {
+                entity.lastUpdatedBy =
+                    context?.requestHeaders?.upn || context?.requestHeaders?.requestingSystemId;
             }
-            entity.lastUpdatedBy = id;
         }
     }
+
+    // @ts-ignore
+    public connection: PolarisConnection;
     public dataVersionHandler: DataVersionHandler;
     public findHandler: FindHandler;
     public softDeleteHandler: SoftDeleteHandler;
+    // @ts-ignore
+    protected repositories: Array<PolarisRepository<any>>;
 
-    constructor(connection: Connection) {
-        super(connection, connection.createQueryRunner());
-        this.dataVersionHandler = new DataVersionHandler(this);
+    constructor(connection: PolarisConnection) {
+        super((connection as unknown) as Connection, connection?.createQueryRunner());
+        this.dataVersionHandler = new DataVersionHandler((this as unknown) as EntityManager);
         this.findHandler = new FindHandler();
-        this.softDeleteHandler = new SoftDeleteHandler(this);
+        this.softDeleteHandler = new SoftDeleteHandler((this as unknown) as EntityManager);
+    }
+
+    // @ts-ignore
+    public getRepository<Entity>(
+        target: (new () => Entity) | Function | EntitySchema<Entity> | string,
+    ): PolarisRepository<Entity> {
+        // throw exception if there is no repository with this target registered
+        if (!this.connection.hasMetadata(target as any)) {
+            throw new RepositoryNotFoundError(this.connection.name, target);
+        }
+        // find already created repository instance and return it if found
+        const metadata = this.connection.getMetadata(target as any);
+        const repository: PolarisRepository<any> | undefined = this.repositories.find(
+            repo => repo.metadata === metadata,
+        );
+        if (repository) {
+            return repository;
+        }
+        // if repository was not found then create it, store its instance and return it
+        const newRepository: PolarisRepository<any> = new PolarisRepositoryFactory().create(
+            (this as unknown) as EntityManager,
+            metadata,
+            this.queryRunner,
+        );
+        this.repositories.push(newRepository);
+        return newRepository;
+    }
+
+    public hasRepository<Entity>(
+        target: (new () => Entity) | Function | EntitySchema<Entity> | string,
+    ): boolean {
+        return this.connection.hasMetadata(target as any);
     }
 
     public async delete<Entity>(
@@ -63,11 +110,7 @@ export class PolarisEntityManager extends EntityManager {
             return this.wrapTransaction(async () => {
                 criteria.context = criteria.context || {};
                 await this.dataVersionHandler.updateDataVersion(criteria.context);
-                const config = this.connection.options.extra.config;
-                if (
-                    (config && config.allowSoftDelete === false) ||
-                    !targetOrEntity.toString().includes(CommonModel.name)
-                ) {
+                if (this.connection.options.extra?.config?.allowSoftDelete === false) {
                     return super.delete(targetOrEntity, criteria.criteria);
                 }
                 return this.softDeleteHandler.softDeleteRecursive(targetOrEntity, criteria);
@@ -85,7 +128,7 @@ export class PolarisEntityManager extends EntityManager {
         if (criteria instanceof PolarisFindOneOptions) {
             return super.findOne(
                 entityClass,
-                this.calculateCriteria<Entity>(entityClass, true, criteria),
+                this.findHandler.findConditions<Entity>(true, criteria),
                 maybeOptions,
             );
         } else {
@@ -98,10 +141,7 @@ export class PolarisEntityManager extends EntityManager {
         criteria?: PolarisFindManyOptions<Entity> | any,
     ): Promise<Entity[]> {
         if (criteria instanceof PolarisFindManyOptions) {
-            return super.find(
-                entityClass,
-                this.calculateCriteria<Entity>(entityClass, true, criteria),
-            );
+            return super.find(entityClass, this.findHandler.findConditions<Entity>(true, criteria));
         } else {
             return super.find(entityClass, criteria);
         }
@@ -114,7 +154,7 @@ export class PolarisEntityManager extends EntityManager {
         if (criteria instanceof PolarisFindManyOptions) {
             return super.count(
                 entityClass,
-                this.calculateCriteria<Entity>(entityClass, false, criteria),
+                this.findHandler.findConditions<Entity>(false, criteria),
             );
         } else {
             return super.count(entityClass, criteria);
@@ -126,10 +166,7 @@ export class PolarisEntityManager extends EntityManager {
         maybeEntityOrOptions?: PolarisSaveOptions<Entity, T> | any,
         maybeOptions?: any,
     ): Promise<T | T[]> {
-        if (
-            maybeEntityOrOptions instanceof PolarisSaveOptions &&
-            targetOrEntity.toString().includes(CommonModel.name)
-        ) {
+        if (maybeEntityOrOptions instanceof PolarisSaveOptions) {
             return this.wrapTransaction(async () => {
                 maybeEntityOrOptions.context = maybeEntityOrOptions.context || {};
                 await this.dataVersionHandler.updateDataVersion(maybeEntityOrOptions.context);
@@ -149,8 +186,9 @@ export class PolarisEntityManager extends EntityManager {
         criteria: PolarisCriteria | any,
         partialEntity: any,
     ): Promise<UpdateResult> {
-        if (criteria instanceof PolarisCriteria) {
-            return this.wrapTransaction(async () => {
+        return this.wrapTransaction(async () => {
+            let updateCriteria = criteria;
+            if (criteria instanceof PolarisCriteria) {
                 criteria.context = criteria.context || {};
                 await this.dataVersionHandler.updateDataVersion(criteria.context);
                 const globalDataVersion = criteria.context.returnedExtensions.globalDataVersion;
@@ -164,11 +202,36 @@ export class PolarisEntityManager extends EntityManager {
                     lastUpdatedBy: upnOrRequestingSystemId,
                 };
                 delete partialEntity.realityId;
-                return super.update(target, criteria.criteria, partialEntity);
+                updateCriteria = criteria.criteria;
+            }
+
+            if (
+                this.connection.options.type === 'postgres' ||
+                this.connection.options.type === 'mssql'
+            ) {
+                return super.update(target, updateCriteria, partialEntity);
+            }
+
+            if (typeof updateCriteria === 'string' || updateCriteria instanceof Array) {
+                updateCriteria = {
+                    where: {
+                        id: In(updateCriteria instanceof Array ? updateCriteria : [updateCriteria]),
+                    },
+                };
+            }
+
+            const entitiesToUpdate = await super.find(target, updateCriteria);
+            entitiesToUpdate.forEach((entityToUpdate: typeof target, index) => {
+                entitiesToUpdate[index] = { ...entityToUpdate, ...partialEntity };
             });
-        } else {
-            return super.update(target, criteria, partialEntity);
-        }
+            await super.save(target, entitiesToUpdate);
+            const updateResult: UpdateResult = {
+                generatedMaps: [],
+                raw: entitiesToUpdate,
+                affected: entitiesToUpdate.length,
+            };
+            return updateResult;
+        });
     }
 
     private async wrapTransaction(action: any) {
@@ -189,11 +252,5 @@ export class PolarisEntityManager extends EntityManager {
             await runner.rollbackTransaction();
             throw err;
         }
-    }
-
-    private calculateCriteria<Entity>(target: any, includeLinkedOper: boolean, criteria: any) {
-        return target.toString().includes(CommonModel.name)
-            ? this.findHandler.findConditions<Entity>(includeLinkedOper, criteria)
-            : criteria;
     }
 }